code_analysis/pyproject.toml
[project]
name = "code-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "amir",email = "amir@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.5"
mypy = "^1.15.0"
ruff = "^0.11.0"
types-requests = "^2.32.0.20250306"

